Basic player damage code for collision detection
Tested with keyboard inputs and is working. I tried testing with
collision for the projectile but it just isn't working with the assets.
I made my own asset and it worked, so maybe I'm missing something.
Anyway, I left a placeholder name so let me know what I can do to fix it
later.