Kill Counter
Modified BasicLevel.unity with UI updates to include a kill counter.
Added some code to the GenericEnemy.cs, TripleShotEnemyScript.cs, and
TrackingEnemy.cs to allow the KillCount script to be called to increment
the number of kills.

KillCount script could probably be moved/integrated into the Player
script.

I've been having issues with Unity lately in getting the controls to
work so uh, this is largely untested. But looking through the code this
should work. Please test this!

Also, added a new tag called KillCount.